+++ LINVAST.Tests/Imperative/Builders/Python/ExpressionTests.cs
@@ -1,3 +1,4 @@
+using System.Linq;
 using LINVAST.Imperative.Builders.Python;
 using LINVAST.Imperative.Nodes;
 using LINVAST.Imperative.Visitors;
@@ -19,8 +20,49 @@ namespace LINVAST.Tests.Imperative.Builders.Python
             Assert.That(this.ParseExpression("None"), Is.TypeOf<NullLitExprNode>());
         }
 
+        [Test]
+        public void LiteralExprBuildsExprNode()
+        {
+            var literal = this.builder.BuildFromSource("42", parser => parser.literal_expr()).As<LitExprNode>();
+
+            Assert.That(literal.Value, Is.EqualTo(42L));
+        }
+
+        [Test]
+        public void BareYieldBuildsYieldExprNode()
+        {
+            var statement = this.ParseStatement("yield\n");
+            var yieldExpr = statement.As<ExprStatNode>().Expression.As<YieldExprNode>();
+
+            Assert.That(yieldExpr.Value, Is.Null);
+            Assert.That(yieldExpr.IsDelegating, Is.False);
+        }
+
+        [Test]
+        public void YieldValueBuildsYieldExprNode()
+        {
+            var statement = this.ParseStatement("yield 1\n");
+            var yieldExpr = statement.As<ExprStatNode>().Expression.As<YieldExprNode>();
+
+            Assert.That(yieldExpr.Value, Is.TypeOf<LitExprNode>());
+            Assert.That(yieldExpr.IsDelegating, Is.False);
+        }
+
+        [Test]
+        public void YieldFromBuildsDelegatingYieldExprNode()
+        {
+            var statement = this.ParseStatement("yield from items\n");
+            var yieldExpr = statement.As<ExprStatNode>().Expression.As<YieldExprNode>();
+
+            Assert.That(yieldExpr.Value, Is.TypeOf<IdNode>());
+            Assert.That(yieldExpr.IsDelegating, Is.True);
+        }
+
         private ExprNode ParseExpression(string source)
             => this.builder.BuildFromSource(source, parser => parser.test()).As<ExprNode>();
+
+        private StatNode ParseStatement(string source)
+            => this.builder.BuildFromSource(source).As<SourceNode>().Children.Single().As<StatNode>();
 
         private void AssertExpressionValue<T>(string source, T expected)
         {
